chore(GitHubIssueButton): require issueBody. Attach environment context.  Make createIssueURL local
